requirements.txt
agate==1.9.1
annotated-types==0.7.0
attrs==25.4.0
babel==2.18.0
backports.tarfile==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
colorama==0.4.6
cryptography==46.0.5
daff==1.4.2
databricks-sdk==0.63.0
databricks-sql-connector==4.0.5
dbt-adapters==1.17.3
dbt-common==1.34.2
dbt-core==1.10.19
dbt-databricks==1.10.19
dbt-extractor==0.6.0
dbt-protos==1.0.440
dbt-semantic-interfaces==0.9.0
dbt-spark==1.9.3
deepdiff==8.6.1
et_xmlfile==2.0.0
google-auth==2.49.1
idna==3.11
importlib_metadata==8.7.1
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.1.1
jaraco.functools==4.4.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.7.0
leather==0.4.1
lz4==4.4.5
MarkupSafe==3.0.3
mashumaro==3.14
more-itertools==10.8.0
msgpack==1.1.2
networkx==3.2.1
numpy==2.0.2
oauthlib==3.3.1
openpyxl==3.1.5
orderly-set==5.5.0
packaging==26.0
pandas==2.2.3
parsedatetime==2.6
pathspec==0.12.1
protobuf==6.33.5
pyarrow==21.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
snowplow-tracker==1.1.0
sqlparams==6.2.0
sqlparse==0.5.4
text-unidecode==1.3
thrift==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
zipp==3.23.0
